+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 ## Added
-- A quick command line interface to look the enumeration of a permutation class.
+- A quick command line interface to compute the enumeration of a permutation class.
 
 ## 1.4.2 - 2020-06-17
 ### Fixed