+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## Added
+- A quick command line interface to compute the enumeration of a permutation class.
 
 ## 1.4.2 - 2020-06-17
 ### Fixed