+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 1.5.0 - 2020-06-23
 ## Added
 - A quick command line interface to compute the enumeration of a permutation class.
 - `Perm.skew_decomposition` and `Perm.sum_decomposition` methods.